move World and Chunk structs to the different directory
